# api/openapi.yaml
openapi: 3.1.0
info:
  title: ConsorcioAbierto API
  version: 0.1.0
  description: |
    API del SaaS multi-tenant de administración de consorcios.
    Contrato de Fase 0. Convenciones: JSON snake_case, fechas ISO 8601,
    importes como enteros `*_cents` + `currency`, IDs UUID.
    Errores RFC 9457 (application/problem+json).
servers:
  - url: /api/v1
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-ca_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Obligatorio en confirmar, publicar, cobranzas, asignaciones y webhooks.
...
    ConsorcioId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CobranzaId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: Error RFC 9457.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Money:
      type: object
      additionalProperties: false
      required: [amount_cents, currency]
      properties:
        amount_cents:
          type: integer
          format: int64
          description: Importe en centavos (nunca float).
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          default: ARS
    ErrorField:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
    Problem:
      type: object
      required: [type, title, status, code, detail, instance, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
          format: uuid
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorField"
    Page:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          required: [request_id]
          properties:
            request_id:
              type: string
              format: uuid
            page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer
...
    Cobranza:
      type: object
      required: [id, unidad_id, fecha, importe, estado]
      properties:
        id: { type: string, format: uuid }
        unidad_id: { type: string, format: uuid }
        fecha: { type: string, format: date }
        canal: { type: string }
        importe: { $ref: "#/components/schemas/Money" }
        referencia: { type: [string, "null"] }
        estado:
          type: string
          enum: [pendiente_revision, acreditado, rechazado, revertido]
        saldo_a_favor_cents: { type: integer, format: int64 }
    MercadoPagoCobranza:
      type: object
      required: [cobranza, checkout_url, provider, preference_id]
      properties:
        cobranza:
          $ref: "#/components/schemas/Cobranza"
        checkout_url: { type: string, format: uri }
        provider: { type: string, enum: [mercadopago, mock] }
        preference_id: { type: string }
...
paths:
...
  /consorcios/{id}/cobranzas:
    parameters:
      - { $ref: "#/components/parameters/ConsorcioId" }
    get:
      operationId: listCobranzas
      tags: [cobranzas]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
    post:
      operationId: createCobranza
      tags: [cobranzas]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [unidad_id, fecha, importe]
              properties:
                unidad_id: { type: string, format: uuid }
                fecha: { type: string, format: date }
        canal:
          type: string
          enum: [efectivo, transferencia, deposito, tarjeta, cajero, mercadopago, otros]
        importe: { $ref: "#/components/schemas/Money" }
        referencia: { type: [string, "null"] }
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cobranza"
        "409":
          description: Referencia duplicada; requiere resolución.
          $ref: "#/components/responses/Problem"
  /consorcios/{id}/cobranzas/mercado-pago:
    parameters:
      - { $ref: "#/components/parameters/ConsorcioId" }
    post:
      operationId: createMercadoPagoCobranza
      tags: [cobranzas]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [unidad_id, fecha, importe]
              properties:
                unidad_id: { type: string, format: uuid }
                fecha: { type: string, format: date }
                importe: { $ref: "#/components/schemas/Money" }
                referencia: { type: [string, "null"] }
      responses:
        "201":
          description: Cobranza creada con enlace de checkout de Mercado Pago.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MercadoPagoCobranza"
        "409": { $ref: "#/components/responses/Problem" }
...
  /cobranzas/{id}/acreditar:
    parameters:
      - { $ref: "#/components/parameters/CobranzaId" }
    post:
      operationId: acreditarCobranza
      tags: [cobranzas]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                asignaciones:
                  description: Opcional; si se omite, el backend propone FIFO.
                  type: array
                  items:
                    type: object
                    required: [charge_id, amount_cents]
                    properties:
                      charge_id: { type: string, format: uuid }
                      amount_cents: { type: integer, format: int64 }
      responses:
        "200":
          description: Pago acreditado con asignaciones y saldo resultante.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cobranza:
                    $ref: "#/components/schemas/Cobranza"
                  asignaciones:
                    type: array
                    items:
                      type: object
                      properties:
                        charge_id: { type: string, format: uuid }
                        amount_cents: { type: integer, format: int64 }
                  saldo_a_favor_cents: { type: integer, format: int64 }
        "409": { $ref: "#/components/responses/Problem" }
  /cobranzas/{id}:
    parameters:
      - { $ref: "#/components/parameters/CobranzaId" }
    get:
      operationId: getCobranza
      tags: [cobranzas]
      responses:
        "200":
          description: Cobranza con detalle de asignaciones y saldo a favor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cobranza:
                    $ref: "#/components/schemas/Cobranza"
                  asignaciones:
                    type: array
                    items:
                      type: object
                      properties:
                        charge_id: { type: string, format: uuid }
                        amount_cents: { type: integer, format: int64 }
                  saldo_a_favor_cents: { type: integer, format: int64 }
        "404": { $ref: "#/components/responses/Problem" }
  /cobranzas/{id}/recibo.pdf:
    parameters:
      - { $ref: "#/components/parameters/CobranzaId" }
    get:
      operationId: getCobranzaRecibo
      tags: [cobranzas]
      responses:
        "200":
          description: PDF descargable del recibo de la cobranza.
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404": { $ref: "#/components/responses/Problem" }
...
  /cobranzas/{id}/asignaciones:
    parameters:
      - { $ref: "#/components/parameters/CobranzaId" }
    post:
      operationId: ajustarAsignaciones
      tags: [cobranzas]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [asignaciones]
              properties:
                asignaciones:
                  type: array
                  items:
                    type: object
                    required: [charge_id, amount_cents]
                    properties:
                      charge_id: { type: string, format: uuid }
                      amount_cents: { type: integer, format: int64 }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  cobranza:
                    $ref: "#/components/schemas/Cobranza"
                  asignaciones:
                    type: array
                    items:
                      type: object
                      properties:
                        charge_id: { type: string, format: uuid }
                        amount_cents: { type: integer, format: int64 }
                  saldo_a_favor_cents: { type: integer, format: int64 }
        "409": { $ref: "#/components/responses/Problem" }
  /cobranzas/{id}/revertir:
    parameters:
      - { $ref: "#/components/parameters/CobranzaId" }
    post:
      operationId: revertirCobranza
      tags: [cobranzas]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Cobranza revertida con asignaciones restauradas.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cobranza:
                    $ref: "#/components/schemas/Cobranza"
                  asignaciones:
                    type: array
                    items:
                      type: object
                      properties:
                        charge_id: { type: string, format: uuid }
                        amount_cents: { type: integer, format: int64 }
                  saldo_a_favor_cents: { type: integer, format: int64 }
        "409": { $ref: "#/components/responses/Problem" }
...